refactor: extract applyAgentEvent, session factory, notification dispatch
- Move applyAgentEvent + AgentEventContext to src/utils/agent/eventDispatch.ts
- Move ActiveSession construction to src/utils/session/sessionFactory.ts
- Move notification action → target mapping to src/utils/notification/dispatch.ts
- App.vue: 935 → 878 lines
