+++ src/utils/agent/eventDispatch.ts
@@ -1,0 +1,66 @@
+// Pure dispatcher: maps an SseEvent into mutations on an ActiveSession
+// via the AgentEventContext adapter. No Vue refs, no component scope.
+
+import type { ActiveSession } from "../../types/session";
+import type { SseEvent } from "../../types/sse";
+import { EVENT_TYPES } from "../../types/events";
+import { findPendingToolCall, toToolCallEntry } from "./toolCalls";
+import {
+  pushErrorMessage,
+  applyTextEvent,
+  applyToolResultToSession,
+} from "../session/sessionHelpers";
+
+export interface AgentEventContext {
+  session: ActiveSession;
+  setCurrentRoleId: (roleId: string) => void;
+  onRoleChange: () => void;
+  refreshRoles: () => Promise<void>;
+  scrollSidebarToBottom: () => void;
+}
+
+export async function applyAgentEvent(
+  event: SseEvent,
+  ctx: AgentEventContext,
+): Promise<void> {
+  const { session } = ctx;
+  switch (event.type) {
+    case EVENT_TYPES.toolCall:
+      session.toolCallHistory.push(toToolCallEntry(event));
+      ctx.scrollSidebarToBottom();
+      return;
+    case EVENT_TYPES.toolCallResult: {
+      const entry = findPendingToolCall(
+        session.toolCallHistory,
+        event.toolUseId,
+      );
+      if (entry) entry.result = event.content;
+      ctx.scrollSidebarToBottom();
+      return;
+    }
+    case EVENT_TYPES.status:
+      session.statusMessage = event.message;
+      return;
+    case EVENT_TYPES.switchRole:
+      setTimeout(() => {
+        ctx.setCurrentRoleId(event.roleId);
+        ctx.onRoleChange();
+      }, 0);
+      return;
+    case EVENT_TYPES.rolesUpdated:
+      await ctx.refreshRoles();
+      return;
+    case EVENT_TYPES.text:
+      applyTextEvent(session, event.message, event.source ?? "assistant");
+      return;
+    case EVENT_TYPES.toolResult:
+      applyToolResultToSession(session, event.result);
+      return;
+    case EVENT_TYPES.error:
+      console.error("[agent] error event:", event.message);
+      pushErrorMessage(session, event.message);
+      return;
+    case EVENT_TYPES.sessionFinished:
+      return;
+  }
+}
